Abstract NoteRecord row reads
