+++ crates/store/src/db/mod.rs
@@ -18,6 +18,7 @@ use miden_objects::{
     utils::Serializable,
 };
 use rusqlite::vtab::array;
+use sql::utils::{column_value_as_u64, read_block_number};
 use tokio::sync::oneshot;
 use tracing::{info, info_span, instrument};
 
@@ -65,6 +66,61 @@ pub struct NoteRecord {
     pub details: Option<Vec<u8>>,
     pub nullifier: Option<Nullifier>,
     pub merkle_path: MerklePath,
+}
+
+impl NoteRecord {
+    /// Columns from the `notes` table ordered to match [`Self::from_row`].
+    const SELECT_COLUMNS: &'static str = "
+            block_num,
+            batch_index,
+            note_index,
+            note_id,
+            note_type,
+            sender,
+            tag,
+            aux,
+            execution_hint,
+            merkle_path,
+            details,
+            nullifier
+    ";
+
+    /// Parses a row from the `notes` table. The sql selection must use [`Self::SELECT_COLUMNS`] to
+    /// ensure ordering is correct.
+    fn from_row(row: &rusqlite::Row<'_>) -> Result<Self> {
+        let block_num = read_block_number(row, 0)?;
+        let note_index = BlockNoteIndex::new(row.get(1)?, row.get(2)?)?;
+        let note_id = row.get_ref(3)?.as_blob()?;
+        let note_id = RpoDigest::read_from_bytes(note_id)?;
+        let note_type = row.get::<_, u8>(4)?.try_into()?;
+        let sender = AccountId::read_from_bytes(row.get_ref(5)?.as_blob()?)?;
+        let tag: u32 = row.get(6)?;
+        let aux: u64 = row.get(7)?;
+        let aux = aux.try_into().map_err(DatabaseError::InvalidFelt)?;
+        let execution_hint = column_value_as_u64(row, 8)?;
+        let merkle_path_data = row.get_ref(9)?.as_blob()?;
+        let merkle_path = MerklePath::read_from_bytes(merkle_path_data)?;
+        let details_data = row.get_ref(10)?.as_blob_or_null()?;
+        let details = details_data.map(<Vec<u8>>::read_from_bytes).transpose()?;
+        let nullifier = row
+            .get_ref(11)?
+            .as_blob_or_null()?
+            .map(Nullifier::read_from_bytes)
+            .transpose()?;
+
+        let metadata =
+            NoteMetadata::new(sender, note_type, tag.into(), execution_hint.try_into()?, aux)?;
+
+        Ok(NoteRecord {
+            block_num,
+            note_index,
+            note_id,
+            metadata,
+            details,
+            nullifier,
+            merkle_path,
+        })
+    }
 }
 
 impl From<NoteRecord> for proto::Note {
